Update go-eCharger API v1 DE.md
Old version made it hard to find out which parameters are read only and which are read/write. Adding this info to the table makes way more sense imho. The old listing in section 3 may be dropped (was missing some variables btw, I marked them in the table with a "?").